Last of Gans theory
I did EVERYTHING possibly to fix this shit. I can't even believe I spent so much time on this garbage. It doesn't work. I repeat: IT DOES NOT WORK. But at least after 3 more hours I understood why it doesn't work. So I guess it's ok and we have to move on.